+++ readium/shared/r2-shared/src/main/java/org/readium/r2/shared/publication/ContentProtection.kt
@@ -10,7 +10,7 @@
 package org.readium.r2.shared.publication
 
 import org.readium.r2.shared.fetcher.Fetcher
-import org.readium.r2.shared.util.File
+import org.readium.r2.shared.publication.asset.PublicationAsset
 import org.readium.r2.shared.publication.services.ContentProtectionService
 import org.readium.r2.shared.util.Try
 
@@ -24,36 +24,36 @@ import org.readium.r2.shared.util.Try
 interface ContentProtection {
 
     /**
-     * Attempts to unlock a potentially protected file.
+     * Attempts to unlock a potentially protected publication asset.
      *
-     * The Streamer will create a leaf [fetcher] for the low-level [file] access (e.g.
-     * ArchiveFetcher for a ZIP archive), to avoid having each Content Protection open the file to
+     * The Streamer will create a leaf [fetcher] for the low-level [asset] access (e.g.
+     * ArchiveFetcher for a ZIP archive), to avoid having each Content Protection open the asset to
      * check if it's protected or not.
      *
      * A publication might be protected in such a way that the package format can't be recognized,
      * in which case the Content Protection will have the responsibility of creating a new leaf
      * [Fetcher].
      *
-     * @return A [ProtectedFile] in case of success, null if the file is not protected by this
-     * technology or a [Publication.OpeningException] if the file can't be successfully opened,
+     * @return A [ProtectedAsset] in case of success, null if the asset is not protected by this
+     * technology or a [Publication.OpeningException] if the asset can't be successfully opened,
      * even in restricted mode.
      */
     suspend fun open(
-        file: File,
+        asset: PublicationAsset,
         fetcher: Fetcher,
         credentials: String?,
         allowUserInteraction: Boolean,
         sender: Any?
-    ): Try<ProtectedFile, Publication.OpeningException>?
+    ): Try<ProtectedAsset, Publication.OpeningException>?
 
     /**
-     * Holds the result of opening a [File] with a [ContentProtection].
+     * Holds the result of opening a [PublicationAsset] with a [ContentProtection].
      *
-     * @property file Protected file which will be provided to the parsers.
-     * In most cases, this will be the file provided to ContentProtection::open(),
+     * @property asset Protected asset which will be provided to the parsers.
+     * In most cases, this will be the asset provided to ContentProtection::open(),
      * but a Content Protection might modify it in some cases:
-     * - If the original file has a media type that can't be recognized by parsers,
-     *   the Content Protection must return a file with the matching unprotected media type.
+     * - If the original asset has a media type that can't be recognized by parsers,
+     *   the Content Protection must return an asset with the matching unprotected media type.
      * - If the Content Protection technology needs to redirect the Streamer to a different file.
      *   For example, this could be used to decrypt a publication to a temporary secure location.
      *
@@ -70,8 +70,8 @@ interface ContentProtection {
      * It can be used to modify the `Manifest`, the root [Fetcher] or the list of service factories
      * of a [Publication].
      */
-    data class ProtectedFile(
-        val file: File,
+    data class ProtectedAsset(
+        val asset: PublicationAsset,
         val fetcher: Fetcher,
         val onCreatePublication: Publication.Builder.() -> Unit = {}
     )